+++ alert-service/database/alerts.go
@@ -1,0 +1,53 @@
+// Copyright (c) 2021 Sorint.lab S.p.A.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+package database
+
+import (
+	"context"
+
+	"github.com/ercole-io/ercole/v2/model"
+	"github.com/ercole-io/ercole/v2/utils"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+// InsertAlert insert the alert in the database
+func (md *MongoDatabase) InsertAlert(alert model.Alert) (*mongo.InsertOneResult, error) {
+	res, err := md.Client.Database(md.Config.Mongodb.DBName).Collection("alerts").InsertOne(context.TODO(), alert)
+	if err != nil {
+		return nil, utils.NewError(err, "DB ERROR")
+	}
+	return res, nil
+}
+
+// ExistNoDataAlertByHost return true if the host has associated a new NO_DATA alert
+func (md *MongoDatabase) ExistNoDataAlertByHost(hostname string) (bool, error) {
+	//Count the number of new NO_DATA alerts associated to the host
+	val, err := md.Client.Database(md.Config.Mongodb.DBName).Collection("alerts").CountDocuments(context.TODO(), bson.M{
+		"alertCode":          model.AlertCodeNoData,
+		"alertStatus":        model.AlertStatusNew,
+		"otherInfo.hostname": hostname,
+	}, &options.CountOptions{
+		Limit: utils.Intptr(1),
+	})
+	if err != nil {
+		return false, utils.NewError(err, "DB ERROR")
+	}
+
+	//Return true if the count > 0
+	return val > 0, nil
+}
